libxc: use correct return type for do_memory_op()
Currently do_memory_op() is returning int, while the hypervisor is
returning long. This will lead to wrong return informations as soon as
e.g. a pfn larger than about 2 billion (8 TB) is returned.

Use the correct long return type instead and correct the functions
expecting a pfn via the return value of do_memory_op().

Acked-by: Wei Liu <wei.liu2@citrix.com>

+++ tools/libxc/xc_domain.c
@@ -828,7 +828,7 @@ int xc_domain_get_tsc_info(xc_interface *xch,
 
 int xc_domain_maximum_gpfn(xc_interface *xch, domid_t domid, xen_pfn_t *gpfns)
 {
-    int rc = do_memory_op(xch, XENMEM_maximum_gpfn, &domid, sizeof(domid));
+    long rc = do_memory_op(xch, XENMEM_maximum_gpfn, &domid, sizeof(domid));
 
     if ( rc >= 0 )
     {
